[TASK] Deprecate /json file
Back in 2018 the new API was launched and used in the core since v9.2.0
for the update checks see https://review.typo3.org/56468. Therefor the
release information for versions 9 and higher is not longer required to
be provided by this file and removed. Instead a hint to the new API is
delivered and also the root node includes a general hint now. This safes
a lot of bandwidth, about 50 percent is generated by this old file.