[BUGFIX] Fix major version detection for package signatures (#425)
Comparing a version like `"12.4.8"` with `7` will result
in a string comparsion instead of the (intended) float/int
comparison. That is because "12.4.8" can not be implicitly
converted to a float as it contains more than one dot.

Note that "12.4.8" >= 7 is thus equal to `(bool)strcmp("12.4.8", "7")`
which itself yields to false as "1" is compared to "7" and is lower.