[TASK] Update ddev installation instruction for TYPO3 >=12 (#451)
ddev recieved support to regonize changed configuration
file places and nameing for TYPO3 v12 and newer, using
composer.json input ant the pasing the `Typo3Version`
class version code to determine if a composer based
TYPO3 v12 or higher is given. That works, however it
may be necessary to restart ddev after installing our
beloved CMS which can be achieved by executing

```terminal
ddev restart
```

after the setup using cli commands. Without this, the
`additional.php` may not be written by ddev when user
tries to open the backend leading to errors realted
to missing the correct `trustedHostsPattern` needed
for `ddev`. In the past this worked without issues,
because silent upgrade migration has been run even
within the setup and installation process. Usually,
for new installation that does not make sense and
has been removed from TYPO3 for v13 and v14 with [1].

With TYPO3 v13.0 a major feature has been intruduced
which allows to configure the backend url instead of
having it fixed to `/typo3/`. That required bigger
changes and combining frontend and backend entry points
and changed to webserver configuration. For `Apache2`
this has been added and is shipped with the TYPO3
version - NGNIX and other webserver needs adjustment.

The default webserver for ddev is NGNIX and not apache2
and ddev seems not not have the required changes on
board required to satisfy a TYPO3 v13 installation.

That means that it is more than usefull to provide a
little more in the commands which users can simply
copy and paste and having a good kickstart experience
than alternativly trying to explain on `get.typo3.org`
how to change that template manually in ddev.

This change adjusts the ddev based installation
instruction for TYPO3 v12 and v13 to be usable
and working.

[1] https://review.typo3.org/c/Packages/TYPO3.CMS/+/81793