[TASK] Link to dedicated documentation pages for installation method (#490)
It is confusing to users that the "Installation" page only talks about
composer when they are currently trying to do a tar legacy installation.

Also who wants to contribute should be directed to the contribution guide not only to GitHub.

All links are only displayed if the page already existed for that version.

* Apply suggestions from code review

Co-authored-by: Andreas Kienast <akienast@scripting-base.de>