[TASK] Promote the GitLab project template (#488)
* [TASK] Promote the GitLab project template

- Added GitLab as a third deployment option alongside AWS and platform.sh
- Linked directly to the GitLab TYPO3 template and to the related article on typo3.org
- Included GitLab logo image with consistent whitespace for visual alignment
- Improves developer onboarding with GitLab CI/CD integration

* [TASK] Link to dedicated documentation pages for installation method

It is confusing to users that the "Installation" page only talks about
composer when they are currently trying to do a tar legacy installation.

Also who wants to contribute should be directed to the contribution guide not only to GitHub.

All links are only displayed if the page already existed for that version.

* [TASK] Link to dedicated documentation pages for installation method

It is confusing to users that the "Installation" page only talks about
composer when they are currently trying to do a tar legacy installation.

Also who wants to contribute should be directed to the contribution guide not only to GitHub.

All links are only displayed if the page already existed for that version.